+++ soup/JsonObject.cpp
@@ -60,7 +60,7 @@ NAMESPACE_SOUP
 	{
 		if (children.size() <= 0b1111)
 		{
-			uint8_t b = 0b1000'0000 | children.size();
+			uint8_t b = 0b1000'0000 | (uint8_t)children.size();
 			SOUP_RETHROW_FALSE(w.u8(b));
 		}
 		else if (children.size() <= 0xffff)
